netbridge-agent/pyproject.toml
[project]
name = "netbridge-agent"
dynamic = ["version"]
description = "NetBridge agent - system tray app for network tunneling through relay"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.9.0",
    "shared-auth",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "pywin32>=306; sys_platform == 'win32'",
    "azure-storage-blob>=12.19.0",
    "packaging>=24.0",
]

[tool.uv.sources]
shared-auth = { path = "../shared" }

[project.scripts]
netbridge = "netbridge_agent.__main__:main"

[tool.uv]
package = true

[tool.hatch.version]
path = "src/netbridge_agent/__init__.py"

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
    "pytest>=8.0.0",
]
